requirements.txt
celery==4.4.6
Django==3.0.8
djangorestframework==3.11.0
python-decouple==3.3